TYPE: Refactor inlay hints
* Split parameter and type hints into different packages
* Get rid of RsPlainHint interface which has only one implementation
* Add @Suppress("UnstableApiUsage")
* Cleanup
