+++ src/main/kotlin/org/rust/ide/hints/parameter/RsParameterInfoHandler.kt
@@ -3,7 +3,7 @@
  * found in the LICENSE file.
  */
 
-package org.rust.ide.hints
+package org.rust.ide.hints.parameter
 
 import com.intellij.codeInsight.lookup.LookupElement
 import com.intellij.lang.parameterInfo.ParameterInfoContext
