+++ src/test/kotlin/org/rust/ide/hints/parameter/RsInlayParameterHintsProviderTest.kt
@@ -3,23 +3,25 @@
  * found in the LICENSE file.
  */
 
-package org.rust.ide.hints
+package org.rust.ide.hints.parameter
 
+import com.intellij.codeInsight.daemon.impl.HintRenderer
 import com.intellij.openapi.vfs.VirtualFileFilter
+import org.intellij.lang.annotations.Language
+import org.rust.RsTestBase
 import org.rust.fileTreeFromText
 import org.rust.lang.core.psi.RsMethodCall
 
-class RsInlayParameterHintsProviderTest : RsPlainInlayHintsProviderTestBase() {
-
+class RsInlayParameterHintsProviderTest : RsTestBase() {
     fun `test fn args`() = checkByText("""
         fn foo(arg: u32, arg2: u32) {}
         fn main() { foo(/*hint text="arg:"*/0, /*hint text="arg2:"*/1); }
-    """, enabledHints = RsPlainParameterHint.PARAMETER_HINT)
+    """)
 
     fun `test arg out of bounds`() = checkByText("""
         fn foo(arg: u32) {}
         fn main() { foo(/*hint text="arg:"*/0, 1); }
-    """, enabledHints = RsPlainParameterHint.PARAMETER_HINT)
+    """)
 
     fun `test method args`() = checkByText("""
         struct S;
@@ -30,7 +32,7 @@ class RsInlayParameterHintsProviderTest : RsPlainInlayHintsProviderTestBase() {
             let s = S;
             s.foo(/*hint text="arg:"*/0, /*hint text="arg2:"*/1);
         }
-    """, enabledHints = RsPlainParameterHint.PARAMETER_HINT)
+    """)
 
     fun `test struct fn arg`() = checkByText("""
         struct S;
@@ -41,7 +43,7 @@ class RsInlayParameterHintsProviderTest : RsPlainInlayHintsProviderTestBase() {
             let s = S;
             S::foo(/*hint text="self:"*/s, /*hint text="arg:"*/0);
         }
-    """, enabledHints = RsPlainParameterHint.PARAMETER_HINT)
+    """)
 
     fun `test smart hint same parameter name`() = checkByText("""
         fn foo(arg: u32, arg2: u32) {}
@@ -49,7 +51,7 @@ class RsInlayParameterHintsProviderTest : RsPlainInlayHintsProviderTestBase() {
             let arg = 0;
             foo(arg, /*hint text="arg2:"*/1);
         }
-    """, enabledHints = RsPlainParameterHint.PARAMETER_HINT)
+    """)
 
     fun `test smart hint method start with set`() = checkByText("""
         struct S;
@@ -60,7 +62,7 @@ class RsInlayParameterHintsProviderTest : RsPlainInlayHintsProviderTestBase() {
             let s = S;
             s.set_foo(1);
         }
-    """, enabledHints = RsPlainParameterHint.PARAMETER_HINT)
+    """)
 
     fun `test smart hint self call start with set`() = checkByText("""
         struct S;
@@ -71,7 +73,7 @@ class RsInlayParameterHintsProviderTest : RsPlainInlayHintsProviderTestBase() {
             let s = S;
             S::set_foo(s, 0);
         }
-    """, enabledHints = RsPlainParameterHint.PARAMETER_HINT)
+    """)
 
     fun `test smart hint same function name and single parameter`() = checkByText("""
         fn foo(arg: u32) {}
@@ -79,7 +81,7 @@ class RsInlayParameterHintsProviderTest : RsPlainInlayHintsProviderTestBase() {
             let foo = 0;
             foo(foo);
         }
-    """, enabledHints = RsPlainParameterHint.PARAMETER_HINT)
+    """)
 
     fun `test smart hint parameter name and ref input`() = checkByText("""
         fn foo(arg: &u32) {}
@@ -87,7 +89,7 @@ class RsInlayParameterHintsProviderTest : RsPlainInlayHintsProviderTestBase() {
             let arg = 0;
             foo(&arg);
         }
-    """, enabledHints = RsPlainParameterHint.PARAMETER_HINT)
+    """)
 
     fun `test smart hint same method name and single parameter`() = checkByText("""
         struct S;
@@ -98,7 +100,7 @@ class RsInlayParameterHintsProviderTest : RsPlainInlayHintsProviderTestBase() {
             let s = S;
             s.foo(10);
         }
-    """, enabledHints = RsPlainParameterHint.PARAMETER_HINT)
+    """)
 
     fun `test smart hint same method name (self call) and single parameter`() = checkByText("""
         struct S;
@@ -109,41 +111,41 @@ class RsInlayParameterHintsProviderTest : RsPlainInlayHintsProviderTestBase() {
             let s = S;
             S::foo(s, 10);
         }
-    """, enabledHints = RsPlainParameterHint.PARAMETER_HINT)
+    """)
 
     fun `test smart should not annotate tuple structs`() = checkByText("""
         struct TS(i32, f32);
         fn main() {
             let s = TS(5i32, 10.0f32);
         }
-    """, enabledHints = RsPlainParameterHint.PARAMETER_HINT)
+    """)
 
     fun `test fn arg with mut ident`() = checkByText("""
         fn foo(mut arg: u32) {}
         fn main() { foo(/*hint text="arg:"*/0); }
-    """, enabledHints = RsPlainParameterHint.PARAMETER_HINT)
+    """)
 
     fun `test fn arg with mut array`() = checkByText("""
         fn foo([mut x, y]: [i32; 2]) {}
         fn main() { foo(/*hint text="[x, y]:"*/0); }
-    """, enabledHints = RsPlainParameterHint.PARAMETER_HINT)
+    """)
 
     fun `test fn arg with mut tuple`() = checkByText("""
         fn foo((mut x, y): (i32, i32)) {}
         fn main() { foo(/*hint text="(x, y):"*/0); }
-    """, enabledHints = RsPlainParameterHint.PARAMETER_HINT)
+    """)
 
     fun `test fn arg with mut struct`() = checkByText("""
         struct S { x: i32, y: i32 }
         fn foo(S { mut x, y }: S) {}
         fn main() { foo(/*hint text="S {x, y}:"*/0); }
-    """, enabledHints = RsPlainParameterHint.PARAMETER_HINT)
+    """)
 
     fun `test fn arg with mut tuple struct`() = checkByText("""
         struct S(i32, i32);
         fn foo(S(mut x, y): S) {}
         fn main() { foo(/*hint text="S(x, y):"*/0); }
-    """, enabledHints = RsPlainParameterHint.PARAMETER_HINT)
+    """)
 
     fun `test don't touch ast`() {
         fileTreeFromText("""
@@ -164,5 +166,19 @@ class RsInlayParameterHintsProviderTest : RsPlainInlayHintsProviderTestBase() {
         checkAstNotLoaded(VirtualFileFilter.ALL)
         val inlays = handler.getParameterHints(target)
         check(inlays.size == 1)
+    }
+
+    @Suppress("UnstableApiUsage")
+    private fun checkByText(@Language("Rust") code: String) {
+        InlineFile(code.replace(HINT_COMMENT_PATTERN, "<$1/>"))
+
+        RsInlayParameterHints.enabledOption.set(true)
+        RsInlayParameterHints.smartOption.set(true)
+
+        myFixture.testInlays({ (it.renderer as HintRenderer).text }) { it.renderer is HintRenderer }
+    }
+
+    companion object {
+        private val HINT_COMMENT_PATTERN = Regex("""/\*(hint.*?)\*/""")
     }
 }
