+++ src/test/kotlin/org/rust/ide/hints/parameter/RsParameterInfoHandlerTestBase.kt
@@ -3,7 +3,7 @@
  * found in the LICENSE file.
  */
 
-package org.rust.ide.hints
+package org.rust.ide.hints.parameter
 
 import com.intellij.lang.parameterInfo.ParameterInfoHandler
 import com.intellij.psi.PsiElement
