+++ src/test/kotlin/org/rust/ide/hints/type/RsExpressionTypeProviderTest.kt
@@ -3,7 +3,7 @@
  * found in the LICENSE file.
  */
 
-package org.rust.ide.hints
+package org.rust.ide.hints.type
 
 import org.intellij.lang.annotations.Language
 import org.rust.RsTestBase
